% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jquery.R
\name{jquery_core}
\alias{jquery_core}
\title{Obtain jQuery Core as an HTML dependency}
\usage{
jquery_core(major_version = 3, minified = getOption("shiny.minified",
  TRUE))
}
\arguments{
\item{major_version}{The major version of jQuery Core. Currrently 1, 2, and 3 are supported.}

\item{minified}{whether or not to provide a minified JavaScript file.}
}
\value{
A \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} object.
}
\description{
Obtain jQuery Core as an HTML dependency
}
